With Sklearn Library
Lasso Regression (With Library)

This implementation of Lasso Regression uses Python and the sklearn library to fit a linear model to data, while incorporating L1 regularization to help with feature selection. Lasso (Least Absolute Shrinkage and Selection Operator) is a type of regression that penalizes the absolute size of the regression coefficients, encouraging sparsity in the model. This means that it can eliminate some features altogether, leading to simpler, more interpretable models.